Module solana_validator::bootstrap

Structs

GetRpcNodeResult 🔒

Struct to wrap the return value from get_rpc_nodes(). The rpc_contact_info is the peer to download from, and snapshot_hash is the (optional) full and (optional) incremental snapshots to download.

PeerSnapshotHash 🔒

Struct to wrap the peers & snapshot hashes together.

RpcBootstrapConfig

SnapshotHash

A snapshot hash. In this context (bootstrap with incremental snapshots), a snapshot hash is both a full snapshot hash and an (optional) incremental snapshot hash.

Enums

GetRpcNodeError

KnownValidatorsToWaitFor 🔒

When waiting for snapshot hashes from the known validators, should we wait for all or any of them?

Constants

BLACKLIST_CLEAR_THRESHOLD 🔒

If we don’t have any alternative peers after this long, better off trying blacklisted peers again.

GET_RPC_PEERS_TIMEOUT 🔒

If we haven’t found any RPC peers after this time, just give up.

MAX_RPC_CONNECTIONS_EVALUATED_PER_ITERATION

NEWER_SNAPSHOT_THRESHOLD 🔒

If we can’t find a good snapshot download candidate after this time, just give up.

WAIT_FOR_ALL_KNOWN_VALIDATORS 🔒

When downloading snapshots, wait at most this long for snapshot hashes from all known validators. Afterwards, wait for snapshot hashes from any known validator.

Functions

attempt_download_genesis_and_snapshot

build_known_snapshot_hashes 🔒

Build the known snapshot hashes from a set of nodes.

check_vote_account 🔒

do_known_validators_have_all_snapshot_hashes 🔒

Check if we can discover snapshot hashes for the known validators.

download_snapshot 🔒

Download a snapshot

download_snapshots 🔒

Check to see if we can use our local snapshots, otherwise download newer ones.

fail_rpc_node

get_eligible_peer_snapshot_hashes 🔒

Get snapshot hashes from all the eligible peers. This fn will get only one snapshot hash per peer (the one with the highest slot). This may be just a full snapshot hash, or a combo full (i.e. base) snapshot hash and incremental snapshot hash.

get_highest_full_snapshot_hash_for_peer 🔒

Get the highest full snapshot hash for a peer from CRDS

get_highest_incremental_snapshot_hash_for_peer 🔒

Get the highest incremental snapshot hash for a peer from CRDS

get_highest_local_snapshot_hash 🔒

Get the Slot and Hash of the local snapshot with the highest slot. Can be either a full snapshot or an incremental snapshot.

get_peer_snapshot_hashes 🔒

Get peer snapshot hashes

get_rpc_nodes 🔒

Get RPC peer node candidates to download from.

get_rpc_peers 🔒

get_snapshot_hashes_from_known_validators 🔒

Get the snapshot hashes from known validators.

get_vetted_rpc_nodes 🔒

is_known_validator 🔒

retain_peer_snapshot_hashes_that_match_known_snapshot_hashes 🔒

Retain the peer snapshot hashes that match a snapshot hash from the known snapshot hashes

retain_peer_snapshot_hashes_with_highest_full_snapshot_slot 🔒

Retain the peer snapshot hashes with the highest full snapshot slot

retain_peer_snapshot_hashes_with_highest_incremental_snapshot_slot 🔒

Retain the peer snapshot hashes with the highest incremental snapshot slot

rpc_bootstrap

should_use_local_snapshot 🔒

Check to see if bootstrap should load from its local snapshots or not. If not, then snapshots will be downloaded.

shutdown_gossip_service 🔒

start_gossip_node 🔒

verify_reachable_ports 🔒

Type Definitions

KnownSnapshotHashes 🔒

Map full snapshot hashes to a set of incremental snapshot hashes. Each full snapshot hash is treated as the base for its set of incremental snapshot hashes.